+++ Buffs/Frozen.cs
@@ -1,11 +1,5 @@
 ﻿using glacial_inferno.Dusts;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Terraria;
-using Terraria.ID;
 using Terraria.ModLoader;
 
 namespace glacial_inferno.Buffs
