+++ Items/Ammo/Arrows/IceArrow.cs
@@ -1,9 +1,4 @@
 ﻿using glacial_inferno.Projectiles.Ammo.Arrows;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
