+++ Items/Ammo/Bullets/FlamingBullet.cs
@@ -1,14 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Terraria;
+﻿using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
-using System.Net.Cache;
 
-namespace glacial_inferno.Items.Ammo
+namespace glacial_inferno.Items.Ammo.Bullets
 {
     internal class FlamingBullet : ModItem
     {
