+++ Items/Ammo/Bullets/FrozenBullet.cs
@@ -1,21 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Terraria;
+﻿using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
 using glacial_inferno.Buffs;
-using System.Net.Cache;
 
-namespace glacial_inferno.Items.Ammo
+namespace glacial_inferno.Items.Ammo.Bullets
 {
     internal class FrozenBullet : ModItem
     {
         public override void SetDefaults()
         {
-          
+
             Item.width = 30;
             Item.height = 30;
             Item.ammo = AmmoID.Bullet;
@@ -23,17 +17,17 @@ namespace glacial_inferno.Items.Ammo
             Item.consumable = true;
             Item.damage = 10;
             Item.knockBack = 2;
-            Item.shoot =ModContent.ProjectileType<Projectiles.Other.FrozenBullet>();
+            Item.shoot = ModContent.ProjectileType<Projectiles.Other.FrozenBullet>();
             Item.shootSpeed = 10;
             Item.value = 30;
         }
 
         public override void OnHitNPC(Player player, NPC target, NPC.HitInfo hit, int damageDone)
         {
- 
+
             int buffType = ModContent.BuffType<Frozen>();
-            
-            
+
+
             base.OnHitNPC(player, target, hit, damageDone);
             target.AddBuff(buffType, 600);
         }
@@ -46,7 +40,7 @@ namespace glacial_inferno.Items.Ammo
 
             recipe1.AddTile(TileID.Anvils);
             recipe1.ReplaceResult(ModContent.ItemType<FrozenBullet>(), 20);
-       
+
             recipe1.Register();
         }
     }
