+++ Items/Ammo/Other/IcySnowball.cs
@@ -1,18 +1,17 @@
-using System.Numerics;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
-using glacial_inferno.Projectiles.Ammo;
+using glacial_inferno.Projectiles.Ammo.Other;
 
-namespace glacial_inferno.Items.Ammo
+namespace glacial_inferno.Items.Ammo.Other
 {
-    public class HeavySnowball : ModItem
+    public class IcySnowball : ModItem
     {
         // The Display Name and Tooltip of this item can be edited in the Localization/en-US_Mods.glacial_inferno.hjson file.
 
         public override void SetDefaults()
         {
-            Item.damage = 16;
+            Item.damage = 8;
             Item.DamageType = DamageClass.Ranged;
             Item.noUseGraphic = true;
             Item.ammo = AmmoID.Snowball;
@@ -20,19 +19,19 @@ namespace glacial_inferno.Items.Ammo
             Item.noMelee = true;
             Item.useTime = Item.useAnimation = 19;
             Item.useStyle = ItemUseStyleID.Swing;
-            Item.knockBack = 9f;
+            Item.knockBack = 5.75f;
             Item.rare = ItemRarityID.White;
             Item.UseSound = SoundID.Item1;
-            Item.shootSpeed = 4.25f;
-            Item.shoot = ModContent.ProjectileType<HeavySnowballProjectile>();
+            Item.shootSpeed = 7;
+            Item.shoot = ModContent.ProjectileType<IcySnowballProjectile>();
             Item.maxStack = 9999;
         }
 
         public override void AddRecipes()
         {
             Recipe recipe = CreateRecipe(10);
-            recipe.AddRecipeGroup(RecipeGroupID.IronBar, 4);
-            recipe.AddIngredient(ItemID.SnowBlock, 8);
+            recipe.AddIngredient(ItemID.IceBlock, 4);
+            recipe.AddIngredient(ItemID.SnowBlock, 4);
             recipe.AddTile(TileID.WorkBenches);
             recipe.Register();
         }
