+++ Items/Weapons/Melee/EarlyIceSword.cs
@@ -1,7 +1,4 @@
 using glacial_inferno.Projectiles.Weapons.Magic;
-using Microsoft.Xna.Framework;
-using System;
-using System.Collections.Generic;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
