+++ Items/Weapons/Ranged/UpgradedSnowballCannon.cs
@@ -1,4 +1,3 @@
-using System.Numerics;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
