+++ Items/Weapons/ShatterItem.cs
@@ -1,10 +1,6 @@
-﻿using glacial_inferno.Projectiles.Weapons.Magic;
-using Microsoft.Xna.Framework;
+﻿using Microsoft.Xna.Framework;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Terraria;
 using Terraria.ModLoader;
