+++ Projectiles/Ammo/Other/HeavySnowballProjectile.cs
@@ -1,10 +1,8 @@
-// using System.Numerics;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
-using Microsoft.Xna.Framework;
 
-namespace glacial_inferno.Projectiles.Ammo
+namespace glacial_inferno.Projectiles.Ammo.Other
 {
     public class HeavySnowballProjectile : ModProjectile
     {
