+++ Projectiles/Ammo/Other/IcySnowballProjectile.cs
@@ -1,11 +1,9 @@
-// using System.Numerics;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
-using Microsoft.Xna.Framework;
 using glacial_inferno.Buffs;
 
-namespace glacial_inferno.Projectiles.Ammo
+namespace glacial_inferno.Projectiles.Ammo.Other
 {
     public class IcySnowballProjectile : ModProjectile
     {
