+++ Projectiles/Other/FlamingBullet.cs
@@ -1,13 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Terraria;
+﻿using Terraria;
 using Terraria.ModLoader;
 using Terraria.ID;
-using System.Numerics;
-using Microsoft.Xna.Framework;
 
 namespace glacial_inferno.Projectiles.Other
 {
