+++ Projectiles/Weapons/Magic/IceBolt.cs
@@ -1,5 +1,4 @@
-﻿using glacial_inferno.Items.Ammo.Arrows;
-using Terraria;
+﻿using Terraria;
 using Terraria.Audio;
 using Terraria.ID;
 using Terraria.ModLoader;
