+++ Systems/GlobalNPCs.cs
@@ -1,17 +1,11 @@
-﻿using glacial_inferno.Items.Ammo;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using glacial_inferno.Items.Ammo.Bullets;
 using Terraria;
-using Terraria.GameContent;
 using Terraria.ID;
 using Terraria.ModLoader;
 
 namespace glacial_inferno.Systems
 {
-    internal class GlobalNPC_ : GlobalNPC
+    public class GlobalNPC_ : GlobalNPC
     {
         public override void ModifyShop(NPCShop shop)
         {
