+++ glacial_inferno.csproj
@@ -17,23 +17,22 @@
     <PackageReference Include="tModLoader.CodeAssist" Version="0.1.*" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Items\Ammo\Bullets\" />
-    <Folder Include="Items\Armor\" />
     <Folder Include="Items\Accessories\" />
+    <Folder Include="Items\Armor\" />
+    <Folder Include="Items\Potions\" />
     <Folder Include="Items\Vanity\" />
-    <Folder Include="Items\Weapons\Summon\" />
     <Folder Include="Items\Weapons\Ranged\" />
-    <Folder Include="NPCs\" />
     <Folder Include="Buffs\" />
     <Folder Include="Dusts\" />
-    <Folder Include="Systems\" />
+    <Folder Include="Items\Weapons\Summon\" />
     <Folder Include="NPCs\" />
-    <Folder Include="Projectiles\Weapons\Magic\" />
     <Folder Include="Projectiles\Ammo\Bullets\" />
     <Folder Include="Projectiles\Enemy\" />
-    <Folder Include="Projectiles\Other\" />
     <Folder Include="Projectiles\Weapons\Melee\" />
     <Folder Include="Projectiles\Weapons\Ranged\" />
     <Folder Include="Projectiles\Weapons\Summon\" />
+    <Folder Include="Systems\" />
+    <Folder Include="Projectiles\Weapons\Magic\" />
+    <Folder Include="Projectiles\Other\" />
   </ItemGroup>
 </Project>
